Backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
motor==3.6.0
numpy==2.2.1
outcome==1.2.0
pandas==2.2.3
pycparser==2.21
pydantic==2.10.5
pydantic_core==2.27.2
pymongo==4.9.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
selenium==4.14.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.41.3
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.0.6
uvicorn==0.34.0
wsproto==1.2.0
